Update empty list to remove only candidates that dont appear in the new list
