+++ crawler/main.py
@@ -1,6 +1,6 @@
 from login import login
 from get_users import get_all_active_users, separate_by_country
-from set_lists import create_lists, clear_lists, populate_lists
+from set_lists import create_lists, clear_diff_lists, populate_lists
 
 if __name__ == "__main__":
     print("Please login first\n")
@@ -12,6 +12,6 @@ if __name__ == "__main__":
     print("Creating Lists..")
     create_lists(session, countries_dict)
     print("Emptying Lists..")
-    clear_lists(session)
+    clear_diff_lists(session, countries_dict)
     print("Populating Lists..")
     populate_lists(session, countries_dict)
